Replace file cleanup with close status='DELETE'
I overlooked the parameter status (and other parameters) of the close statement.
It was already there with FORTRAN 77: Mea culpa! So many good things so early on, he!
Consequently by suggestion of @jbdv-no after discovery by @markj1069
I am using the builtin facility of close statement to delete files.
It is going to have the consequence that if a test fails due to an actual error,
the testoutput is not going to be deleted.

In a debug-situation I think it is the right thing to leave as much trail as possible,
although originally I was focused on cleaning up no matter what.